server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-pinecone>=0.2.8",
    "pinecone>=6.0.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "werkzeug>=3.1.3",
]
